+++ public/packages/common/src/index.ts
@@ -31,3 +31,4 @@ export * from './key-utils';
 export * from './kill-switch';
 export * from './runtime';
 export * from './manifest-validator';
+export * from './usage-meter';